(workspace-decomposition) Extract shortcuts, registry, prefs, geometry from FlexLayoutWorkspace
Move-only split of the FlexLayoutWorkspace shell into focused modules under
workspace/flexlayout/:

- shortcutSections.js — SHORTCUT_SECTIONS catalog
- ShortcutsHelp.jsx — shortcuts overlay subcomponent (+ its CSS import)
- moduleRegistry.js — MODULE_COMPONENTS + MODULE_TITLES
- uiPreferences.js — applyUIPreferences + useUIPreferences hook
- layoutGeometry.js — tabset geometry + panel move/swap/group helpers as plain
  functions taking (model, layoutRef); the shell wraps them in useCallback

Bodies are unchanged apart from mechanical ctx threading (closure model/layoutRef
become parameters). SHORTCUT_SECTIONS and applyUIPreferences are re-exported from
FlexLayoutWorkspace.jsx so the characterization tests keep their import paths.
